error handling for existing files in server

+++ server.py
@@ -57,23 +57,31 @@ def handle_client(name, conn, addr):
                 break
             elif msg.startswith(UPLOAD_MSG):
                 _, filename, filesize = msg.split(' ')
-                print("Filename:", filename)
-                print("Filesize:", filesize, "bytes")
-
-                send_client(name, "[SERVER GRANT] UPLOAD")
-                filename_loc = "server/" + filename 
                 
-                NUM_CHUNKS = int(filesize) // BUF_SIZE + 1
-                with open(filename_loc,"wb") as file:
-                    for _ in range(NUM_CHUNKS):
-                        chunk = conn.recv(BUF_SIZE)
-                        if not chunk:
-                            break
-                        file.write(chunk)
+                # check if a file with same name alredy exists
+                filename_loc = "server/" + filename
+                file_exists = os.path.exists(filename_loc)
 
-                print(f"File {filename} Received from {name}")
-                send_client(name, f"[SERVER UPDATE] Hi {name}, your file {filename} has been received")
-                send_all(name, f"[SERVER UPDATE] {name} has uploaded file {filename} to the server")
+                if file_exists:
+                    # if another file with same name already exist in the server client can't rewrite it
+                    send_client(name, "[SERVER REJECT] UPLOAD 0, another file with same name exists")
+                    print(f"[SERVER REJECT] UPLOAD from {name} rejected, {filename} already exists")
+                else:
+                    print("Filename:", filename)
+                    print("Filesize:", filesize, "bytes")
+                    send_client(name, "[SERVER GRANT] UPLOAD 1")
+                    NUM_CHUNKS = int(filesize) // BUF_SIZE + 1
+                    with open(filename_loc,"wb") as file:
+                        for _ in range(NUM_CHUNKS):
+                            chunk = conn.recv(BUF_SIZE)
+                            if not chunk:
+                                break
+                            file.write(chunk)
+
+                    print(f"File {filename} Received from {name}")
+                    send_client(name, f"[SERVER UPDATE] Hi {name}, your file {filename} has been received")
+                    send_all(name, f"[SERVER UPDATE] {name} has uploaded file {filename} to the server")
+                    
 
             elif msg.startswith(DOWNLOAD_MSG):
                 _,filename = msg.split(' ')
